requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq
langchain-community>=0.3.0
langchain-mcp-adapters>=0.1.0

# MCP
fastmcp>=2.0.0

# GitHub
PyGithub>=2.3.0

# Memory
mem0ai>=0.1.0

# Utilities
httpx>=0.27.0
beautifulsoup4>=4.12.0
black>=24.0.0
pylint>=3.0.0
sse-starlette>=2.1.0